Float is good enough for our work here, and faster

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/actuators/Motor.java
@@ -42,7 +42,7 @@ public class Motor implements Available {
         return enabled && (motor != null);
     }
 
-    public void setPower(double power) {
+    public void setPower(float power) {
         if (!isAvailable()) {
             return;
         }
